Initial checkin of support for common-prefix completion (Issue #8):
Gut some of the complexity in GeneratorModelWindow; since we sort everything, dynamically pulling in bits isn't needed any more.

CompletionSortProxy is gone; we now have CompletionContext which is sort of 
morphing into the center of analysis for a set of completions.  This is also where we should
move icon generation, etc.

Figure out common prefix in shell, stuff it in completion display (TODO show it).